pulls into module and cleans up functions
