+++ pages/home/content.json
@@ -29,6 +29,7 @@
 	"theteamEmail": "email Sherlock",
 	"theteamEmailLink": "sherlock@bakerstreet.gov",
 	"theteamPhone": "0799786772",
+	"servicesH1": "Services",
 	"services": [
 		"brief",
 		"pre-production",
